+++ projects/libs/colors/src/theme-factory/models/theme.type.ts
@@ -3,21 +3,29 @@
 export interface Theme {
   name: string;
   colors: {
-    'color-text-main': string;
-    'color-text-a': string;
-    'color-text-b': string;
-    'color-text-c': string;
+    white: string;
+    black: string;
 
-    'color-surface-main': string;
+    'color-text': string;
+
     'color-surface-a': string;
     'color-surface-b': string;
     'color-surface-c': string;
+
+    'color-neutral-100': string;
+    'color-neutral-200': string;
+    'color-neutral-300': string;
+    'color-neutral-400': string;
+    'color-neutral': string;
+    'color-neutral-500': string;
+    'color-neutral-600': string;
+    'color-neutral-700': string;
 
     'color-primary-100': string;
     'color-primary-200': string;
     'color-primary-300': string;
     'color-primary-400': string;
-    'color-primary-main': string;
+    'color-primary': string;
     'color-primary-500': string;
     'color-primary-600': string;
     'color-primary-700': string;
@@ -26,25 +34,16 @@ export interface Theme {
     'color-secondary-200': string;
     'color-secondary-300': string;
     'color-secondary-400': string;
-    'color-secondary-main': string;
+    'color-secondary': string;
     'color-secondary-500': string;
     'color-secondary-600': string;
     'color-secondary-700': string;
-
-    'color-tertiary-100': string;
-    'color-tertiary-200': string;
-    'color-tertiary-300': string;
-    'color-tertiary-400': string;
-    'color-tertiary-main': string;
-    'color-tertiary-500': string;
-    'color-tertiary-600': string;
-    'color-tertiary-700': string;
 
     'color-info-100': string;
     'color-info-200': string;
     'color-info-300': string;
     'color-info-400': string;
-    'color-info-main': string;
+    'color-info': string;
     'color-info-500': string;
     'color-info-600': string;
     'color-info-700': string;
@@ -53,7 +52,7 @@ export interface Theme {
     'color-success-200': string;
     'color-success-300': string;
     'color-success-400': string;
-    'color-success-main': string;
+    'color-success': string;
     'color-success-500': string;
     'color-success-600': string;
     'color-success-700': string;
@@ -62,7 +61,7 @@ export interface Theme {
     'color-warning-200': string;
     'color-warning-300': string;
     'color-warning-400': string;
-    'color-warning-main': string;
+    'color-warning': string;
     'color-warning-500': string;
     'color-warning-600': string;
     'color-warning-700': string;
@@ -71,7 +70,7 @@ export interface Theme {
     'color-danger-200': string;
     'color-danger-300': string;
     'color-danger-400': string;
-    'color-danger-main': string;
+    'color-danger': string;
     'color-danger-500': string;
     'color-danger-600': string;
     'color-danger-700': string;